# testsuitegen/sample_applications/api_applications/3_cms_social_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CMS & Social API",
    "description": "A sample API for testing String Constraints, Enums, and Security intents.",
    "version": "1.0.0"
  },
  "paths": {
    "/users/profile": {
      "post": {
        "tags": ["User Profiles"],
        "summary": "Create User Profile",
        "description": "Creates a user profile. \nTests: \n- BOUNDARY_MIN/MAX_LENGTH (username, bio)\n- PATTERN_MISMATCH (username, website)\n- FORMAT_INVALID (email)\n- REQUIRED_FIELD_MISSING (username, email, full_name)",
        "operationId": "create_user_profile_users_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserProfile" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/posts/create": {
      "post": {
        "tags": ["Posts"],
        "summary": "Create Post",
        "description": "Creates a blog post.\nTests:\n- ENUM_MISMATCH (status)\n- BOUNDARY_MIN/MAX_ITEMS (tags)\n- TYPE_VIOLATION (tags must be array)",
        "operationId": "create_post_posts_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BlogPost" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/comments/add": {
      "post": {
        "tags": ["Comments"],
        "summary": "Add Comment",
        "description": "Adds a comment.\nTests:\n- XSS_INJECTION (text)\n- SQL_INJECTION (text)\n- WHITESPACE_ONLY (text)",
        "operationId": "add_comment_comments_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Comment" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/posts/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Search Posts",
        "description": "Search endpoint.\nTests:\n- SQL_INJECTION (q)\n- TYPE_VIOLATION (limit)\n- BOUNDARY_MIN/MAX (limit)",
        "operationId": "search_posts_posts_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Search query string",
              "title": "Q"
            },
            "description": "Search query string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/users/{username}": {
      "get": {
        "tags": ["User Profiles"],
        "summary": "Get User",
        "description": "Get user by username.\nTests:\n- SQL_INJECTION (username - path param)\n- RESOURCE_NOT_FOUND (if we mock 404)",
        "operationId": "get_user_users__username__get",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Username to fetch",
              "title": "Username"
            },
            "description": "Username to fetch"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BlogPost": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 5,
            "title": "Title"
          },
          "content": { "type": "string", "minLength": 10, "title": "Content" },
          "status": {
            "$ref": "#/components/schemas/PostStatus",
            "description": "Post publication status"
          },
          "tags": {
            "items": { "type": "string" },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Tags",
            "description": "List of tags"
          },
          "category": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Category",
            "description": "Category (optional)"
          }
        },
        "type": "object",
        "required": ["title", "content", "status", "tags"],
        "title": "BlogPost"
      },
      "Comment": {
        "properties": {
          "post_id": { "type": "integer", "title": "Post Id" },
          "user_id": { "type": "integer", "title": "User Id" },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Comment text. Vulnerable to XSS/SQLi if not sanitized."
          }
        },
        "type": "object",
        "required": ["post_id", "user_id", "text"],
        "title": "Comment"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PostStatus": {
        "type": "string",
        "enum": ["DRAFT", "PUBLISHED", "ARCHIVED"],
        "title": "PostStatus"
      },
      "UserProfile": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_]+$",
            "title": "Username",
            "description": "Alphanumeric username with underscores."
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Valid email address"
          },
          "full_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Full Name"
          },
          "bio": {
            "anyOf": [
              { "type": "string", "maxLength": 500 },
              { "type": "null" }
            ],
            "title": "Bio",
            "description": "Short bio"
          },
          "website": {
            "anyOf": [
              { "type": "string", "pattern": "^https?://" },
              { "type": "null" }
            ],
            "title": "Website",
            "description": "Website URL starting with http/https"
          }
        },
        "type": "object",
        "required": ["username", "email", "full_name"],
        "title": "UserProfile"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}